im_going_insane™
for the life of me i cannot figure out how to extend this refactored gui to add arrows. ive been at this since noon today and these are literally the only changes ive made that actually improve the code: I cannot figure out how to reference the height or width of a "classbox" and as a result the arrows just don't track movement and distance properly. very very sad commit